+++ src/VarNormalAld.cpp
@@ -155,11 +155,7 @@ Rcpp::List VarNormalAldLogLik(Rcpp::List object) {
 
   const arma::mat loglik = bayests::VarNormalAldSampler().log_likelihood(input, draws);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(loglik, "loglik");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "loglik", Rcpp::wrap(loglik));
 }
 
 /*** R