+++ src/VarNormalGamma.cpp
@@ -158,11 +158,7 @@ Rcpp::List VarNormalGammaForecasts(Rcpp::List object) {
   const bayests::ForecastDraws forecast =
     bayests::VarNormalGammaSampler().forecast(input, draws, reporter);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(draws_to_r(forecast.values), "forecast");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "forecast", Rcpp::wrap(draws_to_r(forecast.values)));
 }
 
 // [[Rcpp::export(.VarNormalGammaLogLik)]]
@@ -173,11 +169,7 @@ Rcpp::List VarNormalGammaLogLik(Rcpp::List object) {
 
   const arma::mat loglik = bayests::VarNormalGammaSampler().log_likelihood(input, draws);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(loglik, "loglik");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "loglik", Rcpp::wrap(loglik));
 }
 
 /*** R