+++ src/VarTvpGamma.cpp
@@ -220,11 +220,7 @@ Rcpp::List VarTvpGammaForecasts(Rcpp::List object) {
   const bayests::ForecastDraws forecast =
     bayests::VarTvpGammaSampler().forecast(input, draws, reporter);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(draws_to_r(forecast.values), "forecast");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "forecast", Rcpp::wrap(draws_to_r(forecast.values)));
 }
 
 // [[Rcpp::export(.VarTvpGammaLogLik)]]
@@ -235,11 +231,7 @@ Rcpp::List VarTvpGammaLogLik(Rcpp::List object) {
 
   const arma::mat loglik = bayests::VarTvpGammaSampler().log_likelihood(input, draws);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(loglik, "loglik");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "loglik", Rcpp::wrap(loglik));
 }
 
 /*** R