+++ src/VarTvpStochvol.cpp
@@ -219,11 +219,7 @@ Rcpp::List VarTvpStochvolForecasts(Rcpp::List object) {
   const bayests::ForecastDraws forecast =
     bayests::VarTvpStochvolSampler().forecast(input, draws, reporter);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(draws_to_r(forecast.values), "forecast");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "forecast", Rcpp::wrap(draws_to_r(forecast.values)));
 }
 
 // [[Rcpp::export(.VarTvpStochvolLogLik)]]
@@ -234,11 +230,7 @@ Rcpp::List VarTvpStochvolLogLik(Rcpp::List object) {
 
   const arma::mat loglik = bayests::VarTvpStochvolSampler().log_likelihood(input, draws);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(loglik, "loglik");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "loglik", Rcpp::wrap(loglik));
 }
 
 /*** R