+++ src/VarTvpWishart.cpp
@@ -180,11 +180,7 @@ Rcpp::List VarTvpWishartForecasts(Rcpp::List object) {
   const bayests::ForecastDraws forecast =
     bayests::VarTvpWishartSampler().forecast(input, draws, reporter);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(draws_to_r(forecast.values), "forecast");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "forecast", Rcpp::wrap(draws_to_r(forecast.values)));
 }
 
 // [[Rcpp::export(.VarTvpWishartLogLik)]]
@@ -195,11 +191,7 @@ Rcpp::List VarTvpWishartLogLik(Rcpp::List object) {
 
   const arma::mat loglik = bayests::VarTvpWishartSampler().log_likelihood(input, draws);
 
-  Rcpp::List posterior = object["posterior"];
-  posterior.push_back(loglik, "loglik");
-  object["posterior"] = posterior;
-
-  return object;
+  return with_posterior_element(object, "loglik", Rcpp::wrap(loglik));
 }
 
 /*** R